requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
qrcode>=7.4.2
openpyxl>=3.1.2
tkcalendar>=1.6.1
opencv-python>=4.8.0
pynput>=1.7.6
python-vlc>=3.0.0
requests>=2.31.0
cryptography>=41.0.0
psutil>=5.9.0
